Major update to emission probabilities - Now uses a metric which accounts for the ratio of matches to mismatches. - Appropriate heuristics for the new metric. - Incorporated a temperature term to control emission sensitivity. - Updated tests accordingly.

+++ src/phasing/phasingdptable.cpp
@@ -336,13 +336,13 @@ void PhasingDPTable::get_super_reads(ReadSet* output_read_set) {
 			pos = input_column_iterator.get_position(i);
 			if (active_alleles.size() > 2) {
 				// This position was not phased. Adding BLANKs
-				superreads.first->addVariant(pos, vector<uint32_t>{}, Entry::BLANK);
-				superreads.second->addVariant(pos, vector<uint32_t>{}, Entry::BLANK);	
+				superreads.first->addVariant(pos, vector<uint8_t>{}, Entry::BLANK);
+				superreads.second->addVariant(pos, vector<uint8_t>{}, Entry::BLANK);	
 			} else {
 				assert (active_alleles.size() == 2);
 				// TODO: compute proper weights based on likelihoods.
-				superreads.first->addVariant(pos, vector<uint32_t>{}, population_alleles.allele0);
-				superreads.second->addVariant(pos, vector<uint32_t>{}, population_alleles.allele1);
+				superreads.first->addVariant(pos, vector<uint8_t>{}, population_alleles.allele0);
+				superreads.second->addVariant(pos, vector<uint8_t>{}, population_alleles.allele1);
 			}
 			++i; // next column
 		}
